+++ src/include/rsys/file.h
@@ -192,7 +192,7 @@ typedef enum { NoIndex,
         savea3 = EM_A3;                           \
         EM_A0 = US_TO_SYN68K(pb);                 \
         EM_D0 = err;                              \
-        CALL_EMULATOR(US_TO_SYN68K((long)compp)); \
+        CALL_EMULATOR(US_TO_SYN68K(compp));       \
         EM_D0 = saved0;                           \
         EM_D1 = saved1;                           \
         EM_D2 = saved2;                           \
