+++ src/xdata.cpp
@@ -380,7 +380,7 @@ xdata_for_raw_data(PixMap *target, xdata_t *x, uint32_t *raw_bits,
 	   * longer long-aligned.
 	   */
             x->raw_pat_bits_mem = NewPtr(height * sizeof(uint32_t) + 3);
-            x->pat_bits = (uint32_t *)(((unsigned long)x->raw_pat_bits_mem
+            x->pat_bits = (uint32_t *)(((uintptr_t)x->raw_pat_bits_mem
                                       + 3)
                                      & ~3);
             for(i = 0; i < height; i++)
@@ -409,7 +409,7 @@ xdata_for_raw_data(PixMap *target, xdata_t *x, uint32_t *raw_bits,
 
             /* Allocate long-aligned memory, as described above. */
             x->raw_pat_bits_mem = NewPtr(x->byte_size + 3);
-            x->pat_bits = (uint32_t *)(((unsigned long)x->raw_pat_bits_mem
+            x->pat_bits = (uint32_t *)(((uintptr_t)x->raw_pat_bits_mem
                                       + 3)
                                      & ~3);
             memcpy(x->pat_bits, raw_bits, x->byte_size);
